pyproject.toml
[project]
name = "fbme_biolib"
version = "0.1.0"
description = "# FBME Bio Library – Tools for Biomedical and Bioinformatics Research"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "biopython>=1.85",
    "numpy>=2.2.3",
]


[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["fbme_biolib"]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.7",
]
